Only return terms with three or more characters
Could be configured higher or lower as needed. Eliminates results such
as 's' or other orphaned characters.

("friend's" would be tokenized and stored as "friend" and "s" because
of the apostrophe. It is also possible to leave the original
apostrophe, which would usually be done on a field intended for
searching so that both would match. However for a top terms field
leaving in punctuation marks such as hyphens or apostrophes is
problematic. This field is intended specifically for aggregating the
top terms, not searching. To that end another optimization might be to
only store unique terms, so that they would only count once per
document.